refactor: move SelectText to shared
